Remove false positive making the instance detected as sending unusual traffic while it's not
I remember to have tested that `str_contains` was working as expected
even if the second argument was an integer and not a string but I guess
I haven't been rigorous this time. Complying with [the documentation](https://www.php.net/manual/en/function.str-contains.php) solves the issue that was receiving `YouTube Data API v3 key added.` while it wasn't.